requirements.txt
viser
opencv-python
imageio
imageio-ffmpeg
matplotlib
tensorboard
scikit-learn
jaxtyping
roma
ninja
scikit-image
scipy
pytorch-msssim
fsspec
loguru
jstyleson
pyvista[all]
pyrender
gradio==5.12.0
einops
beartype 
ujson
fpsample
pydantic==2.10.6
json_tricks
hydra-core
colorama
open3d
git+https://github.com/NVlabs/nvdiffrast
fire
pudb
trimesh
nanomesh
transformations==2025.1.1
mediapy
--extra-index-url https://download.pytorch.org/whl/cu121
torch==2.4.1
torchvision==0.19.1
ipdb
nerfview
torchmetrics
splines
pyyaml
black
isort
--extra-index-url https://pypi.nvidia.com
cudf-cu12
dask-cudf-cu12
cuml-cu12
cugraph-cu12
cuspatial-cu12
cuproj-cu12
cuxfilter-cu12
cucim-cu12
pylibraft-cu12
raft-dask-cu12
cuvs-cu12